// apps/webviews/src/types/theme.ts
export const LAYOUT_KEYS = [
  'default',
  'intro',
  'section',
  'quote',
  'image',
  'video',
  'image-left',
  'image-right',
  'two-columns',
] as const;

export type LayoutKey = (typeof LAYOUT_KEYS)[number];